Remove basic auth from staging site `dev.healthequitytracker.org` (#2604)
# Description and Motivation
<!--- bulleted, high level items. use keywords (eg "closes #144" or
"fixes #4323") -->

Initially in place to ensure users didn't accidentally visit the preview
site, the dev staging site was password protected with basic auth
headers. this is cumbersome, and we are no longer in a flow where things
could be on DEV that aren't PROD ready. We do not currently password
protect the netlify deploy previews; however we need to ensure all of
the non-prod sites have robots.txt or similar to prevent search engine
crawling etc.

Needed for #2589 so that we can run the full nightly tests against dev
anytime a PR is merged to main

## Has this been tested? How?



## Screenshots (if appropriate)

## Types of changes

(leave all that apply)

- Bug fix
- New content or feature
- Refactor / chore

## New frontend preview link is below in the Netlify comment 😎