Finishes Tailwind Refactor (#2722)
# Description and Motivation
<!--- bulleted, high level items. use keywords (eg "closes #144" or
"fixes #4323") -->

Removes SASS and related packages; only remaining CSS is co-located in
`index.css`

- closes #2521 
- closes #2522 
- closes #2524
- closes #2622 

## CSS Overrides
- we needed to adjust the color of the generated `<path>` elements that
D3 makes for the time chart x and y boundary axes; to do so we add a new
class name to the "whitelist" in the eslint config file, define the
class in `index.css`, and then simply add that class to the component's
`className` attribute alongside any Tailwind utility classes

## NPM Packages
- installs needed fonts via npm with font source and imports into
`<App>` (removes from index.html and elsewhere)
- removes sass, stylelint, old unused fonts, and normalize packages

## Responsive Handling for TrendsCharts
- I'd like our responsive breakpoints to always be mobile-first, since
that is more performant on low-power mobile devices, is easier to reason
about if we are always consistent, and that's also how Tailwind works by
default. So i have removed the `max-` classes added in #2705 and swapped
the logic to use normal breakpoint modifiers, two different layouts
conditional on whether it's in compare mode or not. ideally container
queries should be used here once they are supported widely enough (they
are already in tailwind sort of)

## Tweaks
- centers video on default helper box on mobile
- fixes weird territory bubble margins

## Has this been tested? How?

visually. rigorous spot checking and bug finding needs to occur before
we cut a release

## Types of changes

(leave all that apply)

- Bug fix
- Refactor / chore

## New frontend preview link is below in the Netlify comment 😎