More shared python requirements consolidation (#3388)
# Description and Motivation
<!--- bulleted, high level items. use keywords (eg "closes #144" or
"fixes #4323") -->

- update lint rules
- fix some linting errors
- consolidate more python package requirements into parent folder
- for now leaving the packages unversioned, which allows the pip
installation to choose the right version. once we consolidate everything
we can probably add back versioning and keep things unified; and also
maybe integrate the `pip-compile` step from the requirements.in that was
half-implemented previously. it needs to be part of the CI/CD build step
otherwise it'll get lost. we should also look into poetry as a more full
featureed management system for dependencies AND virtual environments


## Has this been tested? How?

- test passing
- no errors or warnings when running `install-all-python` locally
- builds and backend working via `infra-test`

## Types of changes

(leave all that apply)

- Refactor / chore

## New frontend preview link is below in the Netlify comment 😎