populates content on crisis overview and data collection tabs, nav fixes, tooltip addition (#3540)
# Description and Motivation
This PR populates the Crisis Overview, Data Collection, and Addressing
Inequities tabs with relevant content and applies the necessary styling
to the Crisis Overview tab. It also includes edits to the “On This Page”
submenu for better navigation. To enhance mobile usability, fixes were
made to the mobile navigation and submenu sizing to prevent spillover
issues. Additionally, a tooltip feature was added for improved user
guidance.

## Has this been tested? How?
tests passing

## Screenshots (if appropriate)

## Types of changes
- Bug fix
- New content or feature
- Refactor / chore

## New frontend preview link is below in the Netlify comment 😎
add '/policy' to preview link

---------

Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Ben Hammond <benjamin.hammond@gmail.com>
Co-authored-by: Eric M Warren II <eric.m.warren1@gmail.com>