Backend: Fix VERA (#3864)
# Description and Motivation
<!--- bulleted, high level items. use keywords (eg "closes #144" or
"fixes #4323") -->

VERA DAG was causing issues reading the csv from the web. 

- checks in the csv file; not a bad idea in case VERA ever decides to
take down the GitHub file we were hotlinking
- improves the reading of that file directly from `data/` and utilizing
`usecols=` to ignore unused cols per demographic
- updates test files

NOTE: this file is a good candidate for refactoring to polars, as the
tests now take over a minute due to the sheer size. This is tradeoff for
testing the ACTUAL source data rather than a mocked version that is
prone to inconsistencies with the source data and requires upkeep

## Has this been tested? How?

- tests passing
- runs on infra-test

## Screenshots (if appropriate)

## Types of changes

(leave all that apply)

- Refactor / chore

## New frontend preview link is below in the Netlify comment 😎