Joyride react 19 (#4377)
temporarily uses fork of `react-joyride` to be compatible with React 19
(after #4298 the joyride was no longer visible)

we should revert back to using the main package if it is eventually
updated and stays maintained. otherwise we might be better off replacing
this temporary fork with another package entirely that's better
maintained. following issue gilbarbara/react-joyride/issues/1151

also fixes a bug where the titles of the joyride steps where not
rendering white and were basically unreadable